requirements.txt
jupyter == 1.0.0
pylint == 2.3.1
pyspark == 2.4.2
spark-sklearn == 0.3.0
scikit-learn == 0.19.2
numpy == 1.16.3
pandas == 0.24.2
scipy == 1.3.0
